requirements.txt
Flask==2.0.1
black==21.7b0
coveralls==3.2.0
environs==9.3.2
pytest==6.2.4
tox==3.24.1
Flask-MySQL==1.5.2
isort==5.9.3
pytest-cov==2.12.1
flake8==3.9.2